+++ pulse_xmpp_agent/pluginsmachine/plugin_updatenetworkcheck.py
@@ -17,7 +17,7 @@ from lib.agentconffile import (
     rotation_file,
 )
 
-NETWORKVERSION = "5.4.0"
+NETWORKVERSION = "5.4.1"
 
 logger = logging.getLogger()
 plugin = {"VERSION": "2.7", "NAME": "updatenetworkcheck", "TYPE": "machine"}  # fmt: skip